% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{as_list}
\alias{as_list}
\alias{as_list.default}
\title{Convert to a list}
\usage{
as_list(x)

\method{as_list}{default}(x)
}
\arguments{
\item{x}{an object to convert to list}
}
\value{
a list
}
\description{
Convert to a list
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method
}}

