% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{census.at.school.500}
\alias{census.at.school.500}
\title{Census at School 500}
\format{
A data frame with 500 rows and 10 variables:
\describe{
  \item{cellsource}{the source of cellphone money}
  \item{rightfoot}{length of the right foot, mm}
  \item{travel}{travel method used to get to school}
  \item{getlunch}{how they get their lunch}
  \item{height}{height measurement, cm}
  \item{gender}{their biological gender}
  \item{age}{their age, years}
  \item{year}{their school year}
  \item{armspan}{their armspan measurement, cm}
  \item{cellcost}{money spent on cellphones}
}
}
\source{
\url{http://new.censusatschool.org.nz/}
}
\usage{
census.at.school.500
}
\description{
A dataset containing 500 observations from a New Zealand census of school
students.
}
\keyword{datasets}
