% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdocuments.R
\name{data_store}
\alias{data_store}
\alias{data_store.default}
\title{Data Store class generator}
\usage{
data_store(x, data)

\method{data_store}{default}(x, data)
}
\arguments{
\item{x}{a data store constructor}

\item{data}{data to be stored (a data.frame)}
}
\value{
a data_store object
}
\description{
Data Store class generator
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for storing at a local path
}}

