% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{dispatch}
\alias{dispatch}
\alias{dispatch.default}
\title{inzight action dispatch method}
\usage{
dispatch(state, action)

\method{dispatch}{default}(state, action)
}
\arguments{
\item{state}{appliation state (inzstate)}

\item{action}{action to apply (inzaction)}
}
\value{
an inzstate
}
\description{
inzight action dispatch method
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method
}}

