% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdocs.R
\name{docs}
\alias{docs}
\title{inzight document collection constructor}
\usage{
docs(...)
}
\arguments{
\item{...}{list of documents}
}
\value{
an \code{inzdocs} object, which is a list of \code{inzdoc} objects
}
\description{
inzight document collection constructor
}
