% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gap Minder}
\format{
An object of class \code{data.frame} with 3577 rows and 55 columns.
}
\source{
Based on free material from GAPMINDER.ORG, CC-BY LICENSE \url{https://www.gapminder.org/data/}
}
\usage{
gapminder
}
\description{
A dataset containing variables for countries over several years.
You can find out more about the dataset from \url{https://www.gapminder.org/data/}.
}
\keyword{datasets}
