% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gimagebutton.R
\name{gimagebutton}
\alias{gimagebutton}
\title{iNZight Image Button}
\usage{
gimagebutton(stock.id = NULL, filename, old_cursor = NULL, tooltip = NULL, ...)
}
\arguments{
\item{stock.id}{the icon to use}

\item{filename}{the image location}

\item{old_cursor}{the cursor before mouseover}

\item{tooltip}{a list with items `tooltip` and `widget`}

\item{...}{additional arguments passed to `gimage`}
}
\value{
a gimage with extra effects
}
\description{
An image icon that works as a button.
}
\details{
Creates an image that works as a button, and has mouseover icon effects
and a 'tooltip' that appears in the status bar.
}
\author{
Tom Elliott
}
