% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtoolbariNZ.R
\name{.gtoolbariNZ}
\alias{.gtoolbariNZ}
\alias{add.GToolBariNZ}
\alias{svalue<-.GToolBariNZ}
\title{generic for toolkit dispatch}
\usage{
.gtoolbariNZ(toolkit, toolbar.list = list(), style = c("both", "icons",
  "text", "both-horiz"), container = NULL, ...)

\method{add}{GToolBar}(obj, child, expand = FALSE, fill = NULL,
  anchor = NULL, ...)

\method{svalue}{GToolBar} (obj, index=NULL, ...) <- value
}
\description{
generic for toolkit dispatch

A toolbar item is a list of action items or a toolbar instance

for a toolbar, \code{svalue<-} replaces the toolbar items with new ones specified by value.
}

