% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzaction}
\alias{inzaction}
\title{inzight action constructor}
\usage{
inzaction(action, ...)
}
\arguments{
\item{action}{the name of the action}

\item{...}{action payload}
}
\value{
an inzaction object
}
\description{
inzight action constructor
}
