% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzcontrols.R
\name{inzcontrol}
\alias{inzcontrol}
\alias{dispatch.inzcontrol}
\title{inzight control state}
\usage{
inzcontrol(name, options, value = "", slider = NULL)

\method{dispatch}{inzcontrol}(state, action)
}
\arguments{
\item{name}{the name of the control (displayed to users)}

\item{options}{valid options for the control}

\item{value}{the chosen value}

\item{slider}{optional, a slider object associated with the control (inzslider)}
}
\description{
inzight control state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzcontrol
}}

