% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzcontrols.R
\name{inzcontrols}
\alias{inzcontrols}
\alias{dispatch.inzcontrols}
\title{inzight controls state}
\usage{
inzcontrols(controls, variables = character())

\method{dispatch}{inzcontrols}(state, action)
}
\arguments{
\item{controls}{a list of controls}

\item{variables}{variable names in the dataset}
}
\value{
a inzcontrols object
}
\description{
inzight controls state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzcontrols
}}

