% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdoc.R
\name{inzdoc}
\alias{inzdoc}
\title{inzight document constructor}
\usage{
inzdoc(key, name = deparse(substitute(data)))
}
\arguments{
\item{name}{the name of the dataset as displayed to users}

\item{path}{the location of the \code{SQLite} database on the server}
}
\value{
an \code{inzdoc} object
}
\description{
inzight document constructor
}
\examples{
inzdoc(iris, name = "Iris Data")
}
