% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdocuments.R
\name{inzdocument}
\alias{inzdocument}
\alias{dispatch.inzdocument}
\title{inzight document state}
\usage{
inzdocument(
  data,
  name = deparse(substitute(data)),
  label = name,
  settings = inzsettings(),
  controls = inzcontrols(variables = names(data))
)

\method{dispatch}{inzdocument}(state, action)
}
\arguments{
\item{data}{a data.frame}

\item{name}{R object name}

\item{label}{human readable name (spaces etc OK)}

\item{settings}{document specific settings (inzsettings)}

\item{controls}{a list of global/user controls (inzcontrols)}
}
\description{
inzight document state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzdocument
}}

