% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdocuments.R
\name{inzdocuments}
\alias{inzdocuments}
\alias{dispatch.inzdocuments}
\title{inzight documents list state}
\usage{
inzdocuments(docs = list(), active = length(docs))

\method{dispatch}{inzdocuments}(state, action)
}
\arguments{
\item{docs}{a list of documents (...inzdocument)}

\item{active}{integer identifying the currently active document}
}
\value{
a inzdocuments object
}
\description{
inzight documents list state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzdocuments
}}

