% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzgraph.R
\name{inzgraph}
\alias{inzgraph}
\alias{dispatch.inzgraph}
\title{inzight graph state}
\usage{
inzgraph(path = tempfile(fileext = ".png"), dimensions = c(600, 400))

\method{dispatch}{inzgraph}(state, action)
}
\arguments{
\item{path}{location to store the file}

\item{dimensions}{width and height of the plot, in pixels}
}
\value{
an inzgraph object
}
\description{
inzight graph state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzgraph
}}

