% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZGUI.R
\docType{class}
\name{iNZGUI-class}
\alias{iNZGUI}
\alias{iNZGUI-class}
\title{main class that builds the iNZight GUI}
\arguments{
\item{data}{an optional data.frame that is loaded
upon initialisation of the GUI window}
}
\description{
main class that builds the iNZight GUI
}

