% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleManager.R
\docType{class}
\name{iNZModule-class}
\alias{iNZModule-class}
\alias{iNZModule}
\title{iNZight Module class for addons}
\description{
iNZight Module
}
\details{
Provides a basic module for extending new ones
}

\author{
Tom Elliott
}
