% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsettings.R
\name{inzsettings}
\alias{inzsettings}
\title{inzight settings constructor}
\usage{
inzsettings(...)
}
\arguments{
\item{...}{settings to be specified}
}
\value{
an \code{inzsettings} object
}
\description{
inzight settings constructor
}
