% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsettings.R
\name{inzsettings_defaults}
\alias{inzsettings_defaults}
\title{Default settings list}
\usage{
inzsettings_defaults()
}
\arguments{
\item{point_colour}{the colour of points}
}
\description{
Default settings list
}
