% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzstate.R
\name{inzstate}
\alias{inzstate}
\alias{dispatch.inzstate}
\title{inzight state object}
\usage{
inzstate(
  documents = inzdocuments(),
  settings = inzsettings(),
  graph = inzgraph()
)

\method{dispatch}{inzstate}(state, action)
}
\arguments{
\item{documents}{a list of documents (inzdocuments)}

\item{settings}{a list of settings (inzsettings)}

\item{graph}{a graph object (inzgraph)}
}
\value{
an inzstate object
}
\description{
inzight state object
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzstate
}}

