% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleManager.R
\docType{class}
\name{NewModuleManager-class}
\alias{NewModuleManager-class}
\alias{NewModuleManager}
\title{Module Manager for iNZight add-ons}
\description{
Module Manager for iNZight add-ons
}
\section{Fields}{

\describe{
\item{\code{GUI}}{an `iNZGUI` object}

\item{\code{win}}{a `gwindow` object}

\item{\code{m_dir}}{path to the user's module directory, grabbed from `GUI`}
}}

\section{Methods}{

\describe{
\item{\code{refresh_modules()}}{Update list of modules installed}
}}

