% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SURFIncomeSurvey}
\alias{SURFIncomeSurvey}
\title{Gap Minder}
\format{A data frame with 3577 rows and 55 variables:
\describe{
  \item{Personid}{a random unique ID number for each record in the Super SURF}
  \item{Gender}{male and female}
  \item{Qualification}{five categories}
  \item{Age}{25 to 64 years in 5-year age bands}
  \item{Hours}{total usual weekly hours worked from all wages and salary jobs}
  \item{Income}{total usual weekly income from all sources}
  \item{Marital}{marital status}
  \item{Ethnicity}{six categories}
}}
\usage{
SURFIncomeSurvey
}
\description{
A dataset containing variables for countries over several years.
}
\keyword{datasets}

