% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SURFIncomeSurvey_200}
\alias{SURFIncomeSurvey_200}
\title{SURF Income Survey}
\format{
A data frame with 200 rows and 8 variables:
\describe{
  \item{Personid}{a random unique ID number for each record in the Super SURF}
  \item{Gender}{male and female}
  \item{Qualification}{five categories}
  \item{Age}{25 to 64 years in 5-year age bands}
  \item{Hours}{total usual weekly hours worked from all wages and salary jobs}
  \item{Income}{total usual weekly income from all sources}
  \item{Marital}{marital status}
  \item{Ethnicity}{six categories}
}
}
\source{
\url{http://www.stats.govt.nz/tools_and_services/university-students/NZIS-Super-SURF.aspx}
}
\usage{
SURFIncomeSurvey_200
}
\description{
The NZIS is run annually during the June quarter (April to June),
as a supplement to the Household Labour Force Survey (HLFS).
All respondents in the HLFS were asked to participate in the NZIS,
which provides a snapshot of income levels for people and households.
NZIS data gives average weekly income for the June quarter from most sources,
including government transfers, investments, self-employment, and wages and salaries.
}
\details{
This is a subset of that data.
}
\keyword{datasets}
