% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzdoc.R
\name{doc}
\alias{doc}
\title{inzight document constructor}
\usage{
doc(data, name = deparse(substitute(data)))
}
\arguments{
\item{data}{the dataset contained within the document}

\item{name}{the name of the dataset as displayed to users}
}
\value{
an \code{inzdoc} object
}
\description{
inzight document constructor
}
\examples{
doc(iris, name = "Iris Data")
}
