% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{Gap Minder}
\format{A data frame with 3577 rows and 55 variables:
\describe{
  \item{Country}{the country name}
  \item{Region.Geo}{geographical region}
  \item{Continent}{continent name}
  \item{Region}{region name}
}}
\usage{
gapminder
}
\description{
A dataset containing variables for countries over several years.
}
\keyword{datasets}

