% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzcontrols.R
\name{inzcontrols}
\alias{inzcontrols}
\title{inzight controls constructor}
\usage{
inzcontrols(v1 = NULL, v2 = NULL, g1 = NULL, g2 = NULL)
}
\arguments{
\item{v1}{the name of the first variable}

\item{v2}{the name of the second variable}

\item{g1}{the name of the first subset variable}

\item{g2}{the name of the second subset variable}
}
\value{
an \code{inzcontrols} object
}
\description{
inzight controls constructor
}
