% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZGUI.R
\docType{class}
\name{iNZGUI-class}
\alias{iNZGUI-class}
\alias{iNZGUI}
\title{iNZight GUI Class}
\description{
Main class that builds the iNZight GUI
}
\section{Fields}{

\describe{
\item{\code{iNZDocuments}}{A list of documents containing data, plot settings, etc.}

\item{\code{activeDoc}}{The numeric ID of the currently active document}
}}

\section{Methods}{

\describe{
\item{\code{get_data_object(nrow)}}{return dataset or survey design, if it exists}

\item{\code{initializeGui(data = NULL, disposeR = FALSE, addonDir = NULL, show = TRUE)}}{Initiates the GUI}

\item{\code{new_document(data, suffix)}}{Create a new document based on the existing one (`getActiveDoc()`)}

\item{\code{update_document(data)}}{Update the existing document with new data or survey design}
}}

