% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZight.R
\name{iNZight}
\alias{iNZight}
\title{iNZight GUI}
\usage{
iNZight(data = NULL, dispose_fun = NULL, ...)
}
\arguments{
\item{data}{a data frame}

\item{dispose_fun}{called when the iNZight GUI is closed}

\item{...}{arguments passed to \code{dispose_fun}}
}
\value{
invisibly returns the \code{iNZGUI} object
}
\description{
Starts iNZight, optionally with data and a dispose function.
For example, a script called to open iNZight can automatically
exit R once the iNZight session is complete using \code{q()} (see examples).
}
\examples{
\dontrun{
# to have iNZight quit R without saving after the session is over,
# you may use the following:
iNZight(dispose_fun = q, save = "n")
}
}
\author{
Marco Kuper, Tom Elliott
}
