\name{iNZight}
\alias{iNZight}
\title{
The iNZight data analysis system
}
\description{
Loads the iNZight data analysis system
}
\usage{
iNZight()
}
\arguments{
  \item{data}{
	Optional argument to supply a data.frame used on startup
}
\value{
This function is not expected to return a value.
}
\author{
Marco Kuper
}
\examples{
\dontrun{iNZight()}
}

