% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZight.R
\name{iNZight}
\alias{iNZight}
\title{iNZight GUI}
\usage{
iNZight(data = NULL, disposeR = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{disposeR}{logical, if \code{TRUE}, R will quit when iNZight is closed.}
}
\description{
iNZight GUI
}
\details{
Starts iNZight
}
\author{
Marco Kuper
}
