% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzight}
\alias{inzight}
\title{inzight state management}
\usage{
inzight(state, action)
}
\arguments{
\item{state}{current application state (inzstate)}

\item{action}{action to apply (inzaction)}
}
\value{
an inzstate
}
\description{
inzight state management
}
