% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzight}
\alias{inzight}
\title{inzight state reducer}
\usage{
inzight(state = inzstate(), action)
}
\arguments{
\item{state}{an \code{inzstate} object}

\item{action}{an \code{inzaction} object}
}
\value{
an \code{inzstate} object
}
\description{
Reduces a state based on a given action.
}
\examples{
state <- inzstate()
action <- inzaction('LOAD_DATA', data = iris)
dispatch(state, action)
}
