% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzight}
\alias{inzight}
\title{inzight state reducer}
\usage{
inzight(state = inzstate(), action)
}
\arguments{
\item{state}{an \code{inzstate} object}

\item{action}{an \code{inzaction} object}
}
\value{
an \code{inzstate} object
}
\description{
Reduces a state based on a given action.
}
\examples{
app <- inzight()
tf <- tempfile("iris", fileext = ".csv")
write.csv(iris, tf, row.names = FALSE, quote = FALSE)
on.exit(unlink(tf))
action <- inzaction('LOAD_DATA', file = tf)
inzight(app, action)
}
