% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsettings.R
\name{inzsettings}
\alias{inzsettings}
\alias{dispatch.inzsettings}
\title{inzight settings state}
\usage{
inzsettings(settings = list())

\method{dispatch}{inzsettings}(state, action)
}
\arguments{
\item{settings}{a list of settings}
}
\value{
an inzsettings object
}
\description{
inzight settings state
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispatch}: Dispatch method for inzsettings
}}

