% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzstate}
\alias{inzstate}
\title{inzight state generator}
\usage{
inzstate(d = docs(), active_doc = 0L, c = controls())
}
\arguments{
\item{d}{an \code{inzdocs} object}

\item{c}{an \code{inzcontrols} object}
}
\value{
an \code{inzstate} object
}
\description{
inzight state generator
}
