% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzight.R
\name{inzstate}
\alias{inzstate}
\title{inzight state generator}
\usage{
inzstate(
  docs = inzdocs(),
  active = 0L,
  controls = inzcontrols(),
  settings = inzsettings()
)
}
\arguments{
\item{docs}{an \code{inzdocs} object}

\item{active}{integer indicating the document currently being used}

\item{controls}{an \code{inzcontrols} object}

\item{settings}{an \code{inzsettings} object}
}
\value{
an \code{inzstate} object
}
\description{
inzight state generator
}
